move ds2 to thirdparty/ds2
