+++ core/datastore.go
@@ -10,7 +10,7 @@ import (
 
 	config "github.com/jbenet/go-ipfs/repo/config"
 	u "github.com/jbenet/go-ipfs/util"
-	ds2 "github.com/jbenet/go-ipfs/util/datastore2"
+	ds2 "github.com/jbenet/go-ipfs/thirdparty/datastore2"
 	"github.com/jbenet/go-ipfs/util/debugerror"
 )
