+++ util/testutil/datastore.go
@@ -3,7 +3,7 @@ package testutil
 import (
 	"github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore"
 	syncds "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore/sync"
-	ds2 "github.com/jbenet/go-ipfs/util/datastore2"
+	ds2 "github.com/jbenet/go-ipfs/thirdparty/datastore2"
 )
 
 func ThreadSafeCloserMapDatastore() ds2.ThreadSafeDatastoreCloser {
